+++ src/test/java/com/zysn/passwordmanager/model/account/entity/UserAccountTest.java
@@ -11,7 +11,7 @@ import javax.crypto.spec.SecretKeySpec;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import com.zysn.passwordmanager.model.security.config.AlgorithmConfig;
+import com.zysn.passwordmanager.model.security.algorithm.config.AlgorithmConfig;
 
 public class UserAccountTest {
     private UserAccount userAccount;
