+++ src/test/java/com/zysn/passwordmanager/model/security/algorithm/derivation/Argon2Test.java
@@ -10,7 +10,9 @@ import javax.crypto.spec.SecretKeySpec;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import com.zysn.passwordmanager.model.security.config.AlgorithmConfig;
+import com.zysn.passwordmanager.model.security.algorithm.config.AlgorithmConfig;
+import com.zysn.passwordmanager.model.security.algorithm.derivation.api.KeyDerivationAlgorithm;
+import com.zysn.passwordmanager.model.security.algorithm.derivation.impl.Argon2;
 
 public class Argon2Test {
     private KeyDerivationAlgorithm keyDerivationAlgorithm;
