+++ src/test/java/com/zysn/passwordmanager/model/security/algorithm/derivation/BCryptTest.java
@@ -9,7 +9,9 @@ import javax.crypto.spec.SecretKeySpec;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import com.zysn.passwordmanager.model.security.config.AlgorithmConfig;
+import com.zysn.passwordmanager.model.security.algorithm.config.AlgorithmConfig;
+import com.zysn.passwordmanager.model.security.algorithm.derivation.api.KeyDerivationAlgorithm;
+import com.zysn.passwordmanager.model.security.algorithm.derivation.impl.Bcrypt;
 
 public class BCryptTest {
     private KeyDerivationAlgorithm keyDerivationAlgorithm;
@@ -35,7 +37,8 @@ public class BCryptTest {
     void testDeriveKey() {
         SecretKeySpec masterKey = keyDerivationAlgorithm.deriveKey(source, algorithmConfig);
 
-        byte[] masterKeyByte = new byte[] { -43, -44, 54, 52, 108, -27, -46, -42, 123, -125, 84, 77, 54, -47, 37, 22,
+        byte[] masterKeyByte = new byte[] { 
+        -43, -44, 54, 52, 108, -27, -46, -42, 123, -125, 84, 77, 54, -47, 37, 22,
                 117, 104, 108, 32, -21, -114, -42, -91 };
 
         assertArrayEquals(masterKeyByte, masterKey.getEncoded(),
