+++ src/test/java/com/zysn/passwordmanager/model/security/config/AlgorithmConfigTest.java
@@ -13,6 +13,8 @@ import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import com.zysn.passwordmanager.model.security.algorithm.config.AlgorithmConfig;
+
 class AlgorithmConfigTest {
     private AlgorithmConfig algorithmConfig;
     private String algorithmType;
